crev-web/backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
anthropic>=0.39.0
pydantic>=2.0.0
rich>=13.0.0
click>=8.1.0
